+++ app/src/main/java/com/example/fitaf/Exercise.kt
@@ -7,7 +7,7 @@ data class Exercise(
     val name: String = "",
     var sets: Int = 0,
     var reps: Int = 0,
-    var weight: Float = 0f
+    var weight: Int = 0
 ) {
     override fun toString(): String {
         return "$name - $sets x $reps x $weight kg"
